requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
PyYAML
